backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
PyJWT==2.9.0
mysql-connector-python==9.0.0
bcrypt==4.1.3
Werkzeug==3.0.3
pymysql
requests
